replace the old file with more oop version
